% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotSites_RRACH}
\alias{annotSites_RRACH}
\title{RRACH sites annotation D. melanogaster}
\format{
GRanges
}
\usage{
annotSites_RRACH
}
\description{
RRACH sites annotation for User's guide toy example for D. melanogaster
selected genes
}
\keyword{datasets}
