% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bam_sk1}
\alias{bam_sk1}
\title{Genomic Alignments example - Sk1 yeast}
\format{
GAlignmentPairs
}
\usage{
bam_sk1
}
\description{
Paired-end alignment data extracted from the Schwartz et al., 2013, Cell Sk1 yeast strain dataset.
}
\keyword{datasets}
