% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txRanges.R
\name{exonBlockGen}
\alias{exonBlockGen}
\title{Generate exon coordinates block}
\usage{
exonBlockGen(iGene, geneAnnot)
}
\arguments{
\item{iGene}{Gene}

\item{geneAnnot}{Gene annotation}
}
\value{

}
\description{
Generate exon coordinates block
}
