% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{exonGRanges}
\alias{exonGRanges}
\title{Representing gene models as GRanges from imported BEDFILE}
\usage{
exonGRanges(geneAnnot_GR)
}
\arguments{
\item{geneAnnot_GR}{}
}
\value{

}
\description{
Representing gene models as GRanges from imported BEDFILE
}
