% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txRanges.R
\name{fbind}
\alias{fbind}
\title{Catenate factors example function}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{factors A}

\item{b}{factors B}
}
\value{

}
\description{
This is an example function
}
\examples{
fbind(iris$Species[c(1, 51, 101)], PlantGrowth$group[c(1, 11, 21)])
}
