% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genome_sk1}
\alias{genome_sk1}
\title{Genome example - Sk1 yeast}
\format{
DNAStringSet
}
\usage{
genome_sk1
}
\description{
Genomic sequence for chromosomes 4 and 5 of the yeast strain Sk1
}
\keyword{datasets}
