% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias.R
\name{getDumpedAlignments}
\alias{getDumpedAlignments}
\title{Retrieve dumped alignments}
\usage{
getDumpedAlignments()
}
\description{
Alias for GenomicAlignments::getDumpedAlignments. Retrieve ambiguous reads
not included in the loaded object. For more info check:
\code{\link[GenomicAlignments]{findMateAlignment}}
}
