% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hid_aggregate_DTlist}
\alias{hid_aggregate_DTlist}
\title{Merge lists of data.tables}
\usage{
hid_aggregate_DTlist(DTL1, DTL2, colsToAdd, keepAll = TRUE)
}
\arguments{
\item{DTL1}{list List of data.table with gene names. As output of the
\code{\link{tx_coverageDT}}, \code{\link{tx_nucFreqDT}}, and
\code{\link{tx_covNucFreqDT}} functions.}

\item{DTL2}{list List of data.table with gene names.}

\item{colsToAdd}{character. Numeric column(s) to be aggregated (added).}

\item{keepAll}{logical. Set to FALSE for just keeping data.tables which
name's are in both DTL.}
}
\value{
list
}
\description{
Merge lists of data.tables
}
