% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{hlp_cbind3Tabs}
\alias{hlp_cbind3Tabs}
\title{Helper bind three results tables}
\usage{
hlp_cbind3Tabs(gencoorT, tab1, tab2)
}
\arguments{
\item{tab2}{}
}
\value{

}
\description{
Helper bind three results tables
}
