% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hlp_coverageTab_mc}
\alias{hlp_coverageTab_mc}
\title{Calculate coverage table (coverage, ends, starts) for all gene models
without gene coordinates (Multicore)}
\usage{
hlp_coverageTab_mc(x, nCores)
}
\arguments{
\item{x}{CompressedGRangesList. Genomic Ranges list containing genomic
alignments data by gene. Constructed via the tx_reads() or tx_reads_mc()
functions.}

\item{nCores}{integer. Number of cores to use to run function.}
}
\value{
data.table
}
\description{
Calculate coverage table (coverage, ends, starts) for all gene models
without gene coordinates (Multicore)
}
\author{
M.A. Garcia-Campos
}
