% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hlp_genCoorTab}
\alias{hlp_genCoorTab}
\title{Table with genomic and transcriptomic coordinates}
\usage{
hlp_genCoorTab(x, geneAnnot)
}
\arguments{
\item{x}{CompressedGRangesList. Genomic Ranges list containing genomic
alignments data by gene. Constructed via tx_reads().}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed()}
}
\value{
data.table
}
\description{
Table with genomic and transcriptomic coordinates
}
