% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IUPAC_code_2nucs}
\alias{IUPAC_code_2nucs}
\title{IUPAC ambiguity alphabet (2nc)}
\format{
Character vector
}
\usage{
IUPAC_code_2nucs
}
\description{
IUPAC nucleotide ambiguity alphabet for combinations of two nucleotides,
for use mostly by the txtools functions that generate datatables with
nucleotide frequency
}
\keyword{datasets}
