% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IUPAC_code_simpl}
\alias{IUPAC_code_simpl}
\title{Simplified nucleotide alphabet}
\format{
Character vector
}
\usage{
IUPAC_code_simpl
}
\description{
Simplified nucleotide alphabet for use mostly by the txtools functions that
generate datatables with nucleotide frequency
}
\keyword{datasets}
