% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{mm_geneAnnot}
\alias{mm_geneAnnot}
\title{Download gene annotation - Case study # 2}
\usage{
mm_geneAnnot()
}
\value{

}
\description{
Download mm9 gene annotation to temporary file.
Selected genes (toy example).
}
\details{
Source: UCSC genes https://genome.ucsc.edu/cgi-bin/hgTables
}
