% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{sc_faGenome}
\alias{sc_faGenome}
\title{Download genome to temporary file - Case study # 1}
\usage{
sc_faGenome()
}
\value{
character
}
\description{
Sk1 genome. Source Schwartz et al., 2013 https://doi.org/10.1016/j.cell.2013.10.047
}
