% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{sc_geneAnnot}
\alias{sc_geneAnnot}
\title{Download gene annotation - Case study #1}
\usage{
sc_geneAnnot()
}
\value{
character
}
\description{
Download rRNA gene annotation of Saccharomyces cerevisae strain SK1 to temporary file
}
