% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sc_rRNAmods_Taoka}
\alias{sc_rRNAmods_Taoka}
\title{Sk1 rRNA modifications catalogue}
\format{
data.table
}
\usage{
sc_rRNAmods_Taoka
}
\description{
Source: Taoka et al., 2016 (https://doi.org/10.1093/nar/gkw564)
}
\keyword{datasets}
