% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tk_faGenome}
\alias{tk_faGenome}
\title{Download genome to temporary file - Case study # 3}
\usage{
tk_faGenome()
}
\value{

}
\description{
Thermococcus kodakarensis complete genome
Source: https://www.ncbi.nlm.nih.gov/assembly/GCF_000009965.1/
}
