% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tx_add_diffNucToRef}
\alias{tx_add_diffNucToRef}
\title{Add number of nucleotide reads different to the reference genome}
\usage{
tx_add_diffNucToRef(DT)
}
\arguments{
\item{DT}{data.table. A data.table object. See
\code{\link{tx_makeDT_coverage}}, \code{\link{tx_makeDT_nucFreq}} or
\code{\link{tx_makeDT_covNucFreq}} functions.}
}
\value{
data.table
}
\description{
Add a column to DT of the sum of nucleotide frequency different to the
reference sequence counting deletions, without considering 'N's nor inserts
'.' into the calculation.
}
