% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_diffNucToRefRatio}
\alias{tx_add_diffNucToRefRatio}
\title{Add column of Different Nucleotide to reference ratio}
\usage{
tx_add_diffNucToRefRatio(DT, addDiffandTotalCols = FALSE)
}
\arguments{
\item{DT}{data.table}

\item{addDiffandTotalCols}{Set to TRUE to add counts of total nucleotides
read (nucTotal) and different to reference nucleotide (diffToRef) columns.}
}
\value{

}
\description{
Add column of Different Nucleotide to reference ratio, diffToRef and nucTotal columns are required
}
