% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_endRatio1bpUS}
\alias{tx_add_endRatio1bpUS}
\title{Add ends to coverage ratio 1bp up-stream}
\usage{
tx_add_endRatio1bpUS(DT, minCov = 50)
}
\arguments{
\item{DT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}()
or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{minCov}{numeric. Minimum coverage required to output ratio. If coverage
is less then an NA is output in that position. To output all positions assign 0.}
}
\value{
data.table
}
\description{
Add ends to coverage ratio 1bp up-stream
}
