% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_exonNumber}
\alias{tx_add_exonNumber}
\title{Add exon number}
\usage{
tx_add_exonNumber(txDT, geneAnnot, nCores = 1)
}
\arguments{
\item{txDT}{data.table A data.table as generated by the \code{\link{tx_coverageDT}},
\code{\link{tx_nucFreqDT}}, and \code{\link{tx_covNucFreqDT}} functions.}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed().}

\item{nCores}{integer. Number of cores to use to run function. Multicore
capability is not available in Windows OS.}
}
\description{
Add exon number
}
