% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_geneRegion}
\alias{tx_add_geneRegion}
\title{Add gene regions}
\usage{
tx_add_geneRegion(txDT, geneAnnot, nCores = 1)
}
\arguments{
\item{txDT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed().}

\item{nCores}{integer. Number of cores to use to run function. Multicore
capability is not available in Windows OS.}
}
\value{
data.table
}
\description{
Add gene regions
}
