% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_misincCount}
\alias{tx_add_misincCount}
\title{Add counts of nucleotide reads different to the reference sequence}
\usage{
tx_add_misincCount(DT)
}
\arguments{
\item{DT}{data.table. A table as output by the
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}
}
\value{
data.table
}
\description{
Add a column to a txDT of the sum of nucleotide frequency different to the
reference sequence, counting deletions, without considering 'N's nor inserts
'.' into the calculation.
}
