% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_misincorpRateNucSpec}
\alias{tx_add_misincorpRateNucSpec}
\title{Nucleotide specific misincorporation rate}
\usage{
tx_add_misincorpRateNucSpec(DT, minCov = 50, refNuc, misNuc)
}
\arguments{
\item{refNuc}{Reference nucleotide}

\item{misNuc}{Misincorporated nucleotide}
}
\value{

}
\description{
Calculates the misincorporation rate from a \bold{ref}erence nucleotide to an
specific \bold{mis}incorporated nucleotide. For example, when looking for
positions in cytidines that were read as thymine in RNA after bisulphite treatment.
}
