% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_pos}
\alias{tx_add_pos}
\title{Add position names column to DT}
\usage{
tx_add_pos(DT, sep = ":", check_uniq = T)
}
\arguments{
\item{DT}{data.table.}

\item{sep}{character. Separator between gene and txcoor, by deafult colon sign.}

\item{check_uniq}{logical. Set to false to override unique position names check.}
}
\value{
data.table
}
\description{
Adds a column which pastes the name of the gene with its transcript coordinate
creating a unique position identifier, for use in downstream analysis which
requires it. The column is added after the 'txcoor' column. Unique names are
checked as well.
}
\author{
M.A. Garcia-Campos
}
