% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_relTxPos}
\alias{tx_add_relTxPos}
\title{Add relative position in transcript}
\usage{
tx_add_relTxPos(txDT, round_dig = 3)
}
\arguments{
\item{round_dig}{}
}
\value{
txDT
}
\description{
Adds a numeric variable which represents the relative position of the
nucleotide along the transcript in a scale of 0 (start) to 1 (end).
}
