% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_rollingMean}
\alias{tx_add_rollingMean}
\title{Adding rolling mean to txDT}
\usage{
tx_add_rollingMean(
  DT,
  colName,
  winSize,
  newColName = NULL,
  fill = NA,
  align = "center",
  minCov = 20,
  nCores = 1
)
}
\arguments{
\item{DT}{data.table. Output of txtools data.tables with coverage information
as output of \code{\link{tx_coverageDT}} \code{\link{tx_covNucFreqDT}}
functions.}

\item{colName}{character. Column to which rolling mean will be applied}

\item{winSize}{numeric. Size of window}

\item{newColName}{character. Name to be given to the output. If left empty,
will assign a default name, combination of the \code{colName} and \code{winSize} arguments.}

\item{fill}{Either an empty vector (no fill), or a vector (recycled to)
length 3 giving left, middle and right fills. By default NA.}

\item{align}{character. Align windows on the "left", "center" or "right".}

\item{minCov}{numeric. Minimum coverage required to output ratio. If coverage
is less then an NA is output in that position. To output all positions assign 0.}

\item{nCores}{integer. Number of cores to use to run function. Multicore
capability is not available in Windows OS.}
}
\value{

}
\description{
Adding rolling mean to txDT
}
