% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_spliceSitesLogical}
\alias{tx_add_spliceSitesLogical}
\title{Add SpliceSites}
\usage{
tx_add_spliceSitesLogical(txDT, geneAnnot)
}
\arguments{
\item{txDT}{data.table. Output of txtools data.tables with coverage information
as output of \code{\link{tx_coverageDT}} \code{\link{tx_covNucFreqDT}}
functions.}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed()}
}
\value{
txDT
}
\description{
Add SpliceSites
}
