% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_addRefSeqDT}
\alias{tx_addRefSeqDT}
\title{Add reference sequence}
\usage{
tx_addRefSeqDT(DT, fastaGenome, geneAnnot)
}
\arguments{
\item{DT}{data.table. A summarized data.table object. See tx_coverageDT(),
tx_nucFreq() and tx_covNucFreqDT() functions.}

\item{fastaGenome}{list. The full reference genome sequences, as prepackaged
by BSgenome. See ?BSgenome::available.genomes()}

\item{geneAnnot}{GRanges. Gene annotation loaded as a GenomicRanges object,
see tx_load_bed().}
}
\value{

}
\description{
Add reference sequence
}
\author{
M.A. Garcia-Campos
}
