% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_aggregate_DTlist}
\alias{tx_aggregate_DTlist}
\title{Merge lists of data.tables}
\usage{
tx_aggregate_DTlist(DTL1, DTL2, colsToAdd)
}
\arguments{
\item{DTL1}{data.table}

\item{DTL2}{data.table}

\item{colsToAdd}{character. Numeric column(s) to be aggregated (added).}
}
\value{
list
}
\description{
Merge lists of data.tables
}
