% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_combineTxReads}
\alias{tx_combineTxReads}
\title{Combine txreads}
\usage{
tx_combineTxReads(txReadsList)
}
\arguments{
\item{txReadsList}{list. A list of CompressedGRangesList objects generated
via \code{\link{tx_reads}}().}
}
\value{
CompressedGRangesList
}
\description{
Combine transcript reads processed by tx_reads(), objects must be inside
a simple list.
}
