% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_complete_DT}
\alias{tx_complete_DT}
\title{Complete a DT object missing genes}
\usage{
tx_complete_DT(DT, geneAnnot, fastaGenome = NULL, nCores = 1)
}
\arguments{
\item{DT}{data.table}

\item{geneAnnot}{GRanges. Gene annotation, gene models, as loaded using the
\code{\link{tx_load_bed}} function.}

\item{fastaGenome}{DNAStrinSet. Sequences that comprise the genome in which
gene models are embedded, loaded using the \code{\link{tx_load_bed}} function.}

\item{nCores}{numeric}
}
\value{
data.table
}
\description{
A function that adds the genomic and transcriptomic coordinates of genes
with no data in a data.table generated by the \code{\link{tx_coverageDT}},
\code{\link{tx_nucFreqDT}}, and \code{\link{tx_covNucFreqDT}} functions.
}
\details{
This function is meant to equalize datasets which don't contain some genes,
as such the order of the genes is the same as in the GRanges object used as
gene annotation.
}
