% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_coverageTab}
\alias{tx_coverageTab}
\title{Calculate coverage table cov, ends, starts for all gene models}
\usage{
tx_coverageTab(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Calculate coverage table cov, ends, starts for all gene models
}
\author{
Miguel Angel Garcia-Campos
}
