% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_covNucFreqDT}
\alias{tx_covNucFreqDT}
\title{Summarized Coverage & Nucleotide Frequency data.table}
\usage{
tx_covNucFreqDT(x, geneAnnot, simplify_IUPAC = "splitForceInt")
}
\arguments{
\item{simplify_IUPAC}{}
}
\value{
data.table
}
\description{
Summarized Coverage & Nucleotide Frequency data.table
}
