% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_covTab}
\alias{tx_covTab}
\title{Calculate coverage table cov, ends, starts for all gene models}
\usage{
tx_covTab(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Calculate coverage table cov, ends, starts for all gene models
}
