% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_cut_geneAnnotBytxDT}
\alias{tx_cut_geneAnnotBytxDT}
\title{Cut gene annotation by txDT's genes}
\usage{
tx_cut_geneAnnotBytxDT(txDT, geneAnnot)
}
\arguments{
\item{geneAnnot}{}
}
\value{
txDT
}
\description{
Cut gene annotation by txDT's genes
}
