% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_cutEnds_DT}
\alias{tx_cutEnds_DT}
\title{Cutting 5' and 3' ends of data.table using txcoors}
\usage{
tx_cutEnds_DT(DT, cut_5p = 0, cut_3p = 0)
}
\arguments{
\item{DT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{cut_5p}{integer Basepairs to me removed from the start of DT for each gene}

\item{cut_3p}{integer Basepairs to me removed from the end of DT for each gene}
}
\value{
data.table
}
\description{
This function removes the heading (5'UTR) and trailing (3' UTR) portions of
data.tables.
}
