% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_dm3_geneAnnot}
\alias{tx_dm3_geneAnnot}
\title{D. melanogaster gene annotation subset path}
\usage{
tx_dm3_geneAnnot()
}
\value{
character. Path to dm3 gene annotation file
}
\description{
Returns the local path to the in-package BED file containing gene models
for highly expressed genes in the 'Pasilla' experiment.
}
