% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_extend_UTR}
\alias{tx_extend_UTR}
\title{Extending GRanges 5' and 3' UTR blocks}
\usage{
tx_extend_UTR(GR, ext_5p = 0, ext_3p = 0)
}
\arguments{
\item{GR}{GRanges GRanges containing a blocks}

\item{ext_5p}{integer Number of bp for the 5' UTR blocks to be extended}

\item{ext_3p}{integer Number of bp for the 3' UTR blocks to be extended}
}
\value{
GRanges
}
\description{
A function that extends the 5' and/or 3' IRanges contained in the 'blocks'
column of GRanges objects, which define the exon ranges of a transcript.
As those loaded by ´tx_load_bed()´.
}
