% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_filter_max_width}
\alias{tx_filter_max_width}
\title{Filter ranges by a maximum width}
\usage{
tx_filter_max_width(x, thr)
}
\arguments{
\item{thr}{}
}
\value{

}
\description{
Filter ranges by a maximum width
}
\author{
Miguel Angel Garcia-Campos
}
