% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_genCoorTab}
\alias{tx_genCoorTab}
\title{Table with genomic and transcriptomic coordinates}
\usage{
tx_genCoorTab(x, geneAnnot)
}
\arguments{
\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed()}
}
\value{

}
\description{
Table with genomic and transcriptomic coordinates
}
\author{
M.A. Garcia-Campos
}
