% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_genCoorTab_mc}
\alias{tx_genCoorTab_mc}
\title{Table with genomic and transcriptomic coordinates (Multi-core)}
\usage{
tx_genCoorTab_mc(x, geneAnnot, nCores)
}
\arguments{
\item{x}{CompressedGRangesList. Genomic Ranges list containing genomic
alignments data by gene. Constructed via the tx_reads() or tx_reads_mc()
functions.}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed()}

\item{nCores}{integer. Number of cores to use to run function.}
}
\value{
data.table
}
\description{
This function is the multi-core version of tx_genCoorTab(), and is only
available for use in UNIX-like operative systems.
}
\author{
M.A. Garcia-Campos
}
