% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{tx_generatePairedEndFASTQ}
\alias{tx_generatePairedEndFASTQ}
\title{Generate paired-end FASTQ file}
\usage{
tx_generatePairedEndFASTQ(
  genome,
  geneAnnot,
  readLen_1,
  readLen_2,
  insertSize,
  libSize,
  fileName_r1,
  fileName_r2,
  TPM = NULL,
  NB_r = 5,
  NB_mu = 500,
  nCores
)
}
\arguments{
\item{genome}{list. The full reference genome sequences, as loaded by
\code{\link{tx_load_genome}}() or prepackaged by BSgenome, see ?BSgenome::available.genomes}

\item{geneAnnot}{GRanges. Gene annotation as loaded by \code{\link{tx_load_bed}}().}

\item{readLen_1}{integer. Length of simulated reads1}

\item{readLen_2}{integer. Length of simulated reads2}

\item{insertSize}{integer. Length of simulated insert (gap between reads)}

\item{libSize}{integer. Size of simulated FASTQ file}

\item{fileName_r1}{character. Name of output file reads1.}

\item{fileName_r2}{character. Name of output file reads2.}

\item{TPM}{numeric. Proportion of reads per gene in gene annotation.
Overrides negative binomial distribution generation.}

\item{NB_r}{numeric. Target r dispersion parameter. Bigger values of r tend
to generate a normal distribution. See ?stats::rnbinom()}

\item{NB_mu}{numeric. Target mean of the resulting distribution}

\item{nCores}{integer. Number of cores to run the function with. Multicore
capability not available in Windows OS.}
}
\value{
Writes FASTQ files fileName_r1 and fileName_r2
}
\description{
Simulates a paired-end FASTQ files from a genome and a gene annotation.
The distribution of reads is randomly selected following a negative binomial
distribution.
}
