% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{tx_generateSingleEndFASTQ}
\alias{tx_generateSingleEndFASTQ}
\title{Generate single-end FASTQ file}
\usage{
tx_generateSingleEndFASTQ(
  genome,
  geneAnnot,
  readLen,
  libSize,
  fileName,
  TPM = NULL,
  NB_r = 5,
  NB_mu = 500,
  nCores
)
}
\arguments{
\item{genome}{list. The full reference genome sequences, as loaded by
\code{\link{tx_load_genome}}() or prepackaged by BSgenome, see ?BSgenome::available.genomes}

\item{geneAnnot}{GRanges. Gene annotation as loaded by \code{\link{tx_load_bed}}().}

\item{readLen}{integer. Length of simulated reads}

\item{libSize}{integer. Size of simulated FASTQ file}

\item{fileName}{character. Name of output file.}

\item{TPM}{numeric. Proportion of reads per gene in gene annotation}

\item{NB_r}{numeric. Target r dispersion parameter. Bigger values of r tend
to generate a normal distribution. See ?stats::rnbinom()}

\item{NB_mu}{numeric. Target mean of the resulting distribution}

\item{nCores}{integer. Number of cores to run the function with. Multicore
capability not available in Windows OS.}
}
\value{
Writes FASTQ fileName
}
\description{
Simulates a single-end FASTQ file from a genome and a gene annotation.
The distribution of reads is randomly selected following a negative binomial
distribution.
}
