% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_get_flankSequence}
\alias{tx_get_flankSequence}
\title{Get flanking sequences}
\usage{
tx_get_flankSequence(DT, logi_col, upFlank, doFlank, addNames = FALSE)
}
\arguments{
\item{DT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{logi_col}{character. Name of column of logical class, which indicates queried sites}

\item{upFlank}{numeric. Up-stream flank length}

\item{doFlank}{numeric. Down-stream flank length}

\item{addNames}{logical. Set to TRUE to add names for each sequence in format
"gene:txcoor". Default is FALSE.}
}
\value{
character
}
\description{
Get flanking sequences
}
