% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_get_flanksFromLogicAnnot}
\alias{tx_get_flanksFromLogicAnnot}
\title{Get data from a position and their neighboring positions}
\usage{
tx_get_flanksFromLogicAnnot(DT, annotCol, valuesCol, flankSize, minCov, nCores)
}
\arguments{
\item{DT}{data.table. Output of txtools data.tables with coverage information
as output of \code{\link{tx_coverageDT}} \code{\link{tx_covNucFreqDT}}
functions.}

\item{annotCol}{character. Name of column from which to retrieve centers}

\item{valuesCol}{character. Name of column from which to get values}

\item{flankSize}{numeric. The size of flanks for both left and right side.}

\item{minCov}{numeric. Minimum coverage required to output results of center
position.}

\item{nCores}{integer. Number of cores to use to run function. Multi-core
capability not available in Windows OS.}
}
\value{
matrix
}
\description{
Get the data from positions marked with TRUE in a logical variable as the
center and the values in its vicinity, delimited by a flank size.
Useful to aggregate data surrounding specific sites as motif locations.
}
