% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_get_geneLengths}
\alias{tx_get_geneLengths}
\title{Get length of genes}
\usage{
tx_get_geneLengths(DT)
}
\arguments{
\item{DT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}
}
\value{
named numeric. Length of genes as per times they appear in DT
}
\description{
Outputs the length of the transcripts in the DT
}
