% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_get_metageExons}
\alias{tx_get_metageExons}
\title{Get metagene by exons}
\usage{
tx_get_metageExons(
  txDT,
  colVars,
  nBins,
  geneAnnot = NULL,
  rm_NArows = TRUE,
  nCores = 1
)
}
\arguments{
\item{nCores}{}
}
\value{
matrix
}
\description{
Get metagene by exons
}
