% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_get_metageneRegions}
\alias{tx_get_metageneRegions}
\title{Get metagene regions}
\usage{
tx_get_metageneRegions(
  txDT,
  geneAnnot,
  colVars,
  nBins_5UTR,
  nBins_CDS = NULL,
  nBins_3UTR = NULL,
  nCores = 1
)
}
\arguments{
\item{nCores}{}
}
\description{
Summarizes the values of a column(s) into a selected number of bins per
gene structure region (5'UTR, CDS, 3'UTR).
}
