% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_get_transcriptSeqs}
\alias{tx_get_transcriptSeqs}
\title{Get transcriptome sequences}
\usage{
tx_get_transcriptSeqs(genome, geneAnnot, outFile = NULL, nCores = 1)
}
\arguments{
\item{nCores}{}
}
\value{
character
}
\description{
Get transcriptome sequences
}
