% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_getUnassignedAlignments}
\alias{tx_getUnassignedAlignments}
\title{Retrieve dumped alignments}
\usage{
tx_getUnassignedAlignments()
}
\value{

}
\description{
Retrieve the read mappings that were not assigned to any gene by
\code{\link{tx_reads}}
}
