% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_load_bam}
\alias{tx_load_bam}
\title{Read paired end bam file by yield size}
\usage{
tx_load_bam(
  file,
  yieldSize = "auto",
  scanFlag = "default",
  loadSeq = FALSE,
  recoverDumpedAligns = F,
  verbose = TRUE
)
}
\arguments{
\item{file}{character. Path to the file to read}

\item{yieldSize}{numeric. Number of reads to be processed at a time}

\item{scanFlag}{integer. Flag used to filter reads.
See \code{\link[Rsamtools]{ScanBamParam}}}

\item{loadSeq}{logical. Set to TRUE for loading the sequences contained
in the BAM file}

\item{recoverDumpedAligns}{logical. If set to TRUE ambiguous alignments
will be retrieved and output along correctly paired reads. More info:
?GenomicAlignments::readGAlignments}

\item{verbose}{logical. Set to FALSE to show less information}
}
\value{
GRanges
}
\description{
Reads a file in BAM format by blocks of lines equal to a yield size, either
automatically calculated or specified by the user, and loads it as a
GenomicAlignments object.
}
\examples{
# Loading  in-package BAM file
bamFile <- system.file("extdata", "example_hg19.bam", package = "txtools")
tx_load_bam(bamFile, loadSeq = TRUE, verbose = TRUE)
}
\author{
M.A. Garcia-Campos
}
