% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_load_bed}
\alias{tx_load_bed}
\title{Load gene models from bed-12 and bed-6 files}
\usage{
tx_load_bed(bedfile)
}
\arguments{
\item{bedfile}{character. Gene annotation file in bed-6 or bed-12 format}
}
\value{

}
\description{
Reads and loads bed files checking for compliance with bed file structure
}
\author{
M.A. Garcia-Campos
}
