% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_load_genome}
\alias{tx_load_genome}
\title{Loading genome as DNAStrinSet}
\usage{
tx_load_genome(fastaFile)
}
\arguments{
\item{fastaFile}{path to FASTA format genome sequences}
}
\value{

}
\description{
Loading genome as DNAStrinSet
}
