% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_load_rdsDT}
\alias{tx_load_rdsDT}
\title{Loading RDS files into data.tables}
\usage{
tx_load_rdsDT(file)
}
\arguments{
\item{file}{File name or path to .rds file to be loaded}
}
\value{
data.table
}
\description{
Loading RDS files into data.tables
}
