% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_makeDT_covNucFreq}
\alias{tx_makeDT_covNucFreq}
\alias{tx_covNucFreqDT}
\title{Summarized Coverage & Nucleotide Frequency data.table}
\usage{
tx_makeDT_covNucFreq(
  x,
  geneAnnot,
  genome = NULL,
  simplify_IUPAC = "splitForceInt",
  fullDT = FALSE,
  nCores = 1
)
}
\arguments{
\item{x}{CompressedGRangesList. Genomic Ranges list containing genomic
alignments data by gene. Constructed via tx_reads().}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed()
function.}

\item{genome}{list. The full reference genome sequences, as prepackaged
by BSgenome, See ?BSgenome::available.genomes(); or loaded by \code{\link{tx_load_genome}}}

\item{simplify_IUPAC}{string. Available options are :
\itemize{
\item "not": Will output the complete nucleotide frequency table including
ambiguous reads using the IUPAC ambiguity code. Check: ?Biostrings::IUPAC_CODE_MAP
\item "splitForceInt": Will force an integers split in which ambiguous codes
will be split and assigned half the frequency into their respective nucleotides,
if the frequency is an odd number the uneven count will be assigned as "N".
\item "splitHalf": Ambiguous nucleotide frequencies will be split in half to
their corresponding nucleotides, in cases where frequency is odd creating
non-integer frequencies.
}}

\item{fullDT}{logical. Set to TRUE if it is desired to output a data.table
with all genes and in the same order as 'geneAnnot' object.}

\item{nCores}{integer. Number of cores to use to run function.}
}
\value{
data.table
}
\description{
This function constructs a list of data.tables that contains nucleotide frequency
metrics per nucleotide by transcript:
\itemize{
\item cov = Insert coverage
\item start_5p = read-start counts
\item end_3p = read-end counts
\item A = Adenine
\item C = Cytosine
\item G = Guanine
\item T = Thymine
\item N = Undetermined nucleotide
\item - = Deletion
\item . = Insert, not read gap between read1 and read2
}
The function requires the input of a GRangesList object output by the
\code{\link{tx_reads}} function, which should contain sequence alignments in the
transcriptomic space, and a gene annotation in GRanges format, as loaded by
the \code{\link{tx_load_bed}} function.
}
\details{
This function allows for usage of multiple cores to reduce processing times
in UNIX-like OS.
}
\seealso{
Other makeDT functions: 
\code{\link{tx_makeDT_coverage}()},
\code{\link{tx_makeDT_nucFreq}()}
}
\author{
M.A. Garcia-Campos
}
\concept{makeDT functions}
