% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_merge_DT}
\alias{tx_merge_DT}
\title{Merge data.tables in list to a single data.table}
\usage{
tx_merge_DT(x)
}
\arguments{
\item{x}{list. List of data.tables. A summarized data.table object.
See tx_coverageDT(), tx_nucFreqDT() and tx_covNucFreqDT() functions.}
}
\value{

}
\description{
Merge data.tables in list to a single data.table
}
