% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_nucFreqDT}
\alias{tx_nucFreqDT}
\title{Summarized Nucleotide Frequency data.table}
\usage{
tx_nucFreqDT(x, geneAnnot, simplify_IUPAC = "splitForceInt")
}
\arguments{
\item{x}{CompressedGRangesList. Genomic Ranges list containing genomic
alignments data by gene. Constructed via the tx_reads() or tx_reads_mc()
functions.}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed()}

\item{simplify_IUPAC}{string. Available options are :
\itemize{
\item "not": Will output the complete nucleotide frequency table including
ambiguous reads using the IUPAC ambiguity code.
See: \code{\link[Biostrings]{IUPAC_CODE_MAP}}
\item "splitForceInt": Will force an integers split in which ambiguous codes
will be split and assigned half the frequency into their respective nucleotides,
if the frequency is an odd number the uneven count will be assigned as "N".
\item "splitHalf": Ambiguous nucleotide frequencies will be split in half to
their corresponding nucleotides, in cases where frequency is odd creating
non-integer frequencies.
}}
}
\value{
data.table
}
\description{
Summarized Nucleotide Frequency data.table
}
\author{
M.A. Garcia-Campos
}
