% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_nucFreqTab}
\alias{tx_nucFreqTab}
\title{Calculate nucleotide frequency pileup for all gene models}
\usage{
tx_nucFreqTab(x, simplify_IUPAC = "not")
}
\arguments{
\item{simplify_IUPAC}{}
}
\value{

}
\description{
Calculate nucleotide frequency pileup for all gene models
}
