% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_orderDT}
\alias{tx_orderDT}
\title{Order txDT}
\usage{
tx_orderDT(DT)
}
\arguments{
\item{DT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}
}
\value{
data.table
}
\description{
Orders a txDT by gene name and transcriptomic coordinate (txcoor).
Accordingly reorders chromosome, strand, and gene factors' levels
}
