% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_plots.R
\name{tx_plot_ggseqlogo}
\alias{tx_plot_ggseqlogo}
\title{Plot motif centered in logical annotation}
\usage{
tx_plot_ggseqlogo(DT, logi_col, upFlank, doFlank, method = "bits")
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Plots a motif of the sequence surrounding sites marked as TRUE in a logical
vector in a
}
