% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_plots.R
\name{tx_plot_metageneAtCDS}
\alias{tx_plot_metageneAtCDS}
\title{Plot metagene at CDS}
\usage{
tx_plot_metageneAtCDS(
  txDT,
  geneAnnotation,
  colVars,
  CDS_align,
  upFlank,
  doFlank,
  summ_fun = "sum",
  roll_fun = "sum",
  roll_n = 100,
  roll_align = "center",
  roll_fill = NA,
  smooth = TRUE,
  spar = 0.3,
  na.rm = TRUE,
  normalize = TRUE,
  tick_by = NULL
)
}
\arguments{
\item{roll_n}{numeric. Window size for rolling functions}

\item{roll_align}{character. Either "center" (default), "left" or "right"}

\item{smooth}{logical. Set to FALSE for not smoothing with spline}

\item{spar}{numeric. Smoothing parameter, typically (but not necessarily) in (0,1].
Check stats::smooth.spline}

\item{tick_by}{}
}
\value{
ggplot
}
\description{
Plot metagene at CDS
}
