% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{tx_plot_metaGeneByBins}
\alias{tx_plot_metaGeneByBins}
\title{Plot metagene by bins}
\usage{
tx_plot_metaGeneByBins(
  txDT,
  colName,
  nBins = 100,
  FUN = "mean",
  minTxLength = 300,
  smooth = TRUE,
  na.rm = TRUE
)
}
\arguments{
\item{txDT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{colName}{character. Name of the column of numeric (or logical) class to be plot}

\item{nBins}{integer. Number of bins to divide each gene on}

\item{FUN}{character. Function to be used to summarize values across bins and genes.}

\item{minTxLength}{integer. Minimum transcript length to be included in analysis.}

\item{smooth}{logical. Set to FALSE for not smoothing line.}

\item{na.rm}{logical. Remove NA values from computations, by default is TRUE.}
}
\value{
ggplot
}
\description{
Plot metagene by bins
}
