% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_plots.R
\name{tx_plot_metageneExons}
\alias{tx_plot_metageneExons}
\title{Plot metagene exons}
\usage{
tx_plot_metageneExons(
  txDT,
  colVars,
  nBins,
  geneAnnot = NULL,
  rm_NArows = TRUE,
  nCores = 1,
  summ_fun = "mean",
  smooth = TRUE,
  spar = 0.3,
  plot_type = "lineplot",
  xLabelJump = 10
)
}
\arguments{
\item{xLabelJump}{Number of bins to be skipped for label plotting}
}
\value{
ggplot
}
\description{
Plot metagene exons
}
