% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_plots.R
\name{tx_plot_nucFreq}
\alias{tx_plot_nucFreq}
\title{Nucleotide frequency plot}
\usage{
tx_plot_nucFreq(
  DT,
  txRange = 1:nrow(DT),
  removeInsert = T,
  makePlotly = F,
  show_yLabels = T,
  bar_border = T,
  showLegend = T
)
}
\arguments{
\item{DT}{data.table or data.frame. Input data from which to generate the plot}

\item{txRange}{integer. Range in data to be used, 'txcoor' column is used to
delimit this range in the data.table.}

\item{removeInsert}{logical. Removes counts from insert reads, which do not
call for a nucleotide, are just the gap between read1 and read2.}

\item{makePlotly}{logical. Outputs an interactive plot by using the ggplotly()
function.}

\item{show_yLabels}{logical. If set to FALSE hides the y axis labels.}

\item{bar_border}{logical. If set to FALSE removes the border of the bars.}

\item{showLegend}{logical. If set to FALSE does not renger a legend.}
}
\value{

}
\description{
Generates a ggplot barplot using the nucleotide frequency counts in
data.tables generated by tx_covNucFreqDT(). Automatic Plotly conversion
is available
}
