% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_plots.R
\name{tx_plot_staEndCov}
\alias{tx_plot_staEndCov}
\title{Transcript coverage plot highlighting read-starts and read-ends counts}
\usage{
tx_plot_staEndCov(
  DT,
  gene,
  txRange = 1:nrow(DT),
  makePlotly = FALSE,
  removeCov = FALSE,
  show_yLabels = TRUE,
  bar_border = TRUE,
  showLegend = TRUE
)
}
\arguments{
\item{DT}{data.table or data.frame. Input data from which to generate the plot}

\item{gene}{character. Name of the gene in DT which wants to be plotted.}

\item{txRange}{integer. Range in data to be used, 'txcoor' column is used to
delimit this range in the data.table.}

\item{makePlotly}{logical. Outputs an interactive plot by using the ggplotly()
function.}

\item{removeCov}{logical. If set to TRUE remove coverage counts.}

\item{show_yLabels}{logical. If set to FALSE hides the y axis labels.}

\item{bar_border}{logical. If set to FALSE removes the border of the bars.}

\item{showLegend}{logical. If set to FALSE does not renger a legend.}
}
\value{
ggplot
}
\description{
Transcript coverage plot highlighting read-starts and read-ends counts
}
