% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_reads}
\alias{tx_reads}
\title{Merge paired-end reads and assign them to gene models}
\usage{
tx_reads(
  reads,
  geneAnnot,
  overlapType = "within",
  minReads = 50,
  withSeq = F,
  verbose = T
)
}
\arguments{
\item{reads}{GenomicRanges}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed()}

\item{overlapType}{character. Overlap type to filter reads by gene
model.}

\item{minReads}{integer. Minimum number of reads required to overlap a gene
model to be part of the output object.}

\item{withSeq}{logical. Set to TRUE if sequence should be preserved; 'reads'
object should contain sequences.}

\item{verbose}{logical. Set to FALSE to show less information}
}
\value{
list
}
\description{
Merge paired-end reads and assign them to gene models
}
\author{
M.A. Garcia-Campos
}
