% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_reads_2}
\alias{tx_reads_2}
\title{Transcriptomic reads convertion}
\usage{
tx_reads_2(
  reads,
  geneAnnot,
  minReads = 50,
  withSeq = F,
  verbose = T,
  nCores = 1
)
}
\arguments{
\item{reads}{GAlignments or GAlignmentPairs. Genomic alignments to be processed}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed()}

\item{minReads}{integer. Minimum number of reads required to overlap a gene}

\item{withSeq}{logical. Set to TRUE if sequence should be preserved; 'reads'
object should contain sequences.}

\item{verbose}{logical. Set to FALSE to show less information.}

\item{nCores}{integer. Number of cores to use to run function.}
}
\value{

}
\description{
Assign reads to gene models and convert them to transcriptomic coordinate
system.
}
