% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_split_DT}
\alias{tx_split_DT}
\title{Split data.table to list of data.tables}
\usage{
tx_split_DT(x, dropEmpty = TRUE)
}
\arguments{
\item{x}{data.table. Merged data.table as output by tx_merge_DT()}

\item{dropEmpty}{logical. Drops empty list elements, which occur when data
of genes have been entirely removed, but kept listed in the x$gene factor levels}
}
\value{
list
}
\description{
Split data.table back to list with individual data.tables by 'gene' names
}
