% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_test_ttest}
\alias{tx_test_ttest}
\title{t-test in txDT list}
\usage{
tx_test_ttest(DTL, cont_var, test_groups, test_na.rm = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Apply a t-test per nucleotide position separating by groups over a txDT list.
At least 2 samples in each group are needed.
}
