% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txCoreCols_refSeq}
\alias{txCoreCols_refSeq}
\title{txtools' core cols and refSeq}
\format{
character
}
\usage{
txCoreCols_refSeq
}
\description{
txtools' core column names including refSeq
}
\keyword{datasets}
