% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txRanges.R
\name{txRangeToBED}
\alias{txRangeToBED}
\title{Transcriptomic range to genomic coordinates}
\usage{
txRangeToBED(Tgene, TrangeStart, TrangeEnd, geneAnnot)
}
\arguments{
\item{Tgene}{The gene}

\item{TrangeStart}{txRange start}

\item{TrangeEnd}{txRange end}

\item{geneAnnot}{The gene annotation in BED style}
}
\value{
vector
}
\description{
From transcriptomic range to genomic coordinates in BED format
}
\examples{
# Need to add examples
}
