% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{window_around}
\alias{window_around}
\title{Centered numeric sequence}
\usage{
window_around(position, windowLength)
}
\arguments{
\item{position}{integer. Center of sequence}

\item{windowLength}{integer. Length of both downstream and upstream flanks}
}
\value{
integer
}
\description{
Creates a numerical sequence which is centered in the first argument and
extends in leght of the second argument into both positive and negative
directions.
}
\examples{
# Numeric interval centered in 10, with 5 closest negative
# and positive integers
window_around(10, 5)
}
