% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hlp_genCoorTab_mc}
\alias{hlp_genCoorTab_mc}
\title{Table with genomic and transcriptomic coordinates (Multi-core)}
\usage{
hlp_genCoorTab_mc(x, geneAnnot, nCores)
}
\arguments{
\item{x}{CompressedGRangesList. Genomic Ranges list containing genomic
alignments data by gene. Constructed via tx_reads().}

\item{geneAnnot}{GRanges. Gene annotation as loaded by \code{\link{tx_load_bed}}().}

\item{nCores}{integer. Number of cores to run the function with. Multicore
capability not available in Windows OS.}
}
\value{
data.table
}
\description{
This function is the multi-core version of hlp_genCoorTab(), and is only
available for use in UNIX-like operative systems.
}
