% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_plots.R
\name{hlp_plot_metageneRegions}
\alias{hlp_plot_metageneRegions}
\title{Plot an extracted metaGeneMatrix}
\usage{
hlp_plot_metageneRegions(
  metaGeneMatrix,
  colVars,
  nBins_5UTR,
  nBins_CDS,
  nBins_3UTR,
  summ_fun = "mean",
  smooth = TRUE,
  spar = 0.5,
  plot_type = "lineplot",
  normalize = FALSE
)
}
\arguments{
\item{metaGeneMatrix}{matrix. As obtained by \code{\link{tx_get_metageneRegions}}()}

\item{colVars}{character. Names of columns for which values will be extracted}

\item{nBins_5UTR}{integer. Number of bins into which allocate data on 5'UTR regions}

\item{nBins_CDS}{integer. Number of bins into which allocate data on CDS regions}

\item{nBins_3UTR}{integer. Number of bins into which allocate data on 3'UTR regions}

\item{summ_fun}{character. Summarizing function either "sum" or "mean". Default: "mean"}

\item{smooth}{logical. Set to FALSE for not smoothing line.}

\item{spar}{numeric. Smoothing parameter, typically (but not necessarily) in (0,1].}

\item{plot_type}{character. Type of plot to be output, either "lineplot" or "boxplot".}

\item{normalize}{logical. If set to TRUE, values are normalized so that the
area bellow the curve approximates to 1 for each variable.}
}
\value{
ggplot
}
\description{
It plots an extracted metagene matrix from tx_get_metageneRegions(). Useful
when modifying the metagene matrix, like filtering genes out.
}
\details{
Note It must have the same nBins arguments as the ones used to generate the
metagene matrix.
}
