% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hlp_plot_metageneRegions}
\alias{hlp_plot_metageneRegions}
\title{Plot an extracted metaGeneMatrix}
\usage{
hlp_plot_metageneRegions(
  metaGeneMatrix,
  colVars,
  nBins_5UTR,
  nBins_CDS = NULL,
  nBins_3UTR = NULL,
  summ_fun = "mean",
  smooth = TRUE,
  spar = 0.3,
  plot_type = "lineplot",
  lwd = 1
)
}
\arguments{
\item{plot_type}{character. Either 'lineplot' or 'boxplot'}

\item{lwd}{numeric. Line width for line plot}
}
\description{
It plots an extracted metagene matrix from get_metageneRegions(). Useful
when modifying the metagene matrix, like filtering genes out.
}
\details{
Note It must have the same nBins arguments as the ones used to generate the
metagene matrix.
}
