% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IUPAC_CODE_MAP_extended}
\alias{IUPAC_CODE_MAP_extended}
\title{IUPAC code extended to consider deletions}
\format{
Named character vector
}
\source{
Adapted from Biostrings object \code{\link[Biostrings]{IUPAC_CODE_MAP}}
}
\usage{
IUPAC_CODE_MAP_extended
}
\description{
The IUPAC_CODE_MAP_extended named character vector contains the mapping from
the IUPAC nucleotide ambiguity codes to their meaning adding the deletion
code "-", this is required for computing the consensus strings between
overlapping reads.
}
\keyword{datasets}
