% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tx_add_diffNucToRefRatio}
\alias{tx_add_diffNucToRefRatio}
\title{Add different nucleotide reads to total ratio}
\usage{
tx_add_diffNucToRefRatio(DT, addDiffandTotalCols = FALSE)
}
\arguments{
\item{DT}{data.table}

\item{addDiffandTotalCols}{Set to TRUE to add counts of total nucleotides
read (nucTotal) and different to reference nucleotide (diffToRef) columns.}
}
\value{
data.table
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

tx_add_diffNucToRefRatio() was renamed to tx_add_misincRate()

Add a column to DT of the ratio of different nucleotides to the total of
meaningfuk nucleotide reads not counting undetermined 'N' and inserts '.'.
}
\seealso{
\code{\link{tx_add_diffNucToRef}} and \code{\link{tx_add_nucTotal}}
}
\keyword{internal}
