% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_endRatio}
\alias{tx_add_endRatio}
\title{Add ends to coverage ratio}
\usage{
tx_add_endRatio(DT)
}
\arguments{
\item{DT}{data.table. Output of txtools data.tables with coverage information
as output of \code{\link{tx_coverageDT}} \code{\link{tx_covNucFreqDT}}
functions}
}
\value{
data.table
}
\description{
Adds a column to DT of the read-ends to coverage ratio.
}
