% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_endRatio}
\alias{tx_add_endRatio}
\title{Add ends to coverage ratio 1 bp downstream}
\usage{
tx_add_endRatio(DT, minCov = 50)
}
\arguments{
\item{DT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{minCov}{numeric. Minimum coverage required to output ratio. If coverage
is less then an NA is output in that position.}
}
\value{
data.table
}
\description{
Adds a column to DT of the read-ends to coverage ratio.
}
