% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_exonNumber}
\alias{tx_add_exonNumber}
\title{Add exon number}
\usage{
tx_add_exonNumber(txDT, geneAnnot, nCores = 1)
}
\arguments{
\item{txDT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{geneAnnot}{GRanges. Gene annotation as loaded by \code{\link{tx_load_bed}}().}

\item{nCores}{integer. Number of cores to run the function with. Multicore
capability is not available in Windows OS.}
}
\value{
data.table
}
\description{
Add exon number
}
