% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_exonPlace}
\alias{tx_add_exonPlace}
\title{Add exon place}
\usage{
tx_add_exonPlace(txDT, geneAnnot, nCores = 1)
}
\arguments{
\item{txDT}{data.table A data.table as generated by the \code{\link{tx_coverageDT}},
\code{\link{tx_nucFreqDT}}, and \code{\link{tx_covNucFreqDT}} functions.}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed().}

\item{nCores}{integer. Number of cores to use to run function. Multicore
capability is not available in Windows OS.}
}
\value{
data.table
}
\description{
Adds a column specifying if the exon is the first, intermediate, or last exon
}
