% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_misincRate}
\alias{tx_add_misincRate}
\title{Add misincorporation to total nucleotide reads ratio}
\usage{
tx_add_misincRate(DT, minNucReads = 20, addCounts = FALSE)
}
\arguments{
\item{DT}{data.table. A table as output by the
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{minNucReads}{Minimum number of nucleotides read needed to calculate
misincorporation rate}

\item{addCounts}{Set to TRUE to add counts of total nucleotides
read (nucTotal) and different to reference nucleotide (misincCount) columns.}
}
\value{
data.table
}
\description{
Add a column to txDT of the ratio of different nucleotides to the total of
nucleotide reads, not counting undetermined reads 'N' and inserts '.'.
}
\seealso{
\code{\link{tx_add_diffNucToRef}} and \code{\link{tx_add_nucTotal}}
}
