% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_nucTotal}
\alias{tx_add_nucTotal}
\title{Add total of nucleotide reads per bp}
\usage{
tx_add_nucTotal(DT)
}
\arguments{
\item{DT}{data.table}
}
\value{

}
\description{
Add a column to DT of the sum of nucleotide frequencies for meaningful
nucleotide reads, this is not counting Ns and inserts '.'.
}
