% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_nucTotal}
\alias{tx_add_nucTotal}
\title{Add total number of nucleotide reads}
\usage{
tx_add_nucTotal(DT)
}
\arguments{
\item{DT}{data.table. A table as output by the
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}
}
\value{
data.table
}
\description{
Add a column to DT of the sum of nucleotide frequencies for meaningful
nucleotide reads not counting undetermined 'N' and inserts '.'.
}
