% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_pos}
\alias{tx_add_pos}
\title{Add position unique names}
\usage{
tx_add_pos(DT, sep = ":", check_uniq = T)
}
\arguments{
\item{DT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{sep}{character. Separator between gene and txcoor, by deafult colon sign.}

\item{check_uniq}{logical. Set to false to override unique position names check.}
}
\value{
data.table
}
\description{
Adds a column to DT which pastes the name of the gene with its transcript coordinate
creating unique position identifiers, for use in downstream analysis which
requires them. The column is added after the mandatory 'txcoor' column.
Unique names formation is expected and checked by default.
}
