% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_refSeqDT}
\alias{tx_add_refSeqDT}
\title{Add reference sequence to a data.table}
\usage{
tx_add_refSeqDT(DT, genome, geneAnnot, nCores = 1)
}
\arguments{
\item{DT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{genome}{list. The full reference genome sequences, as loaded by
\code{\link{tx_load_genome}}() or prepackaged by BSgenome, see ?BSgenome::available.genomes}

\item{geneAnnot}{GRanges. Gene annotation as loaded by \code{\link{tx_load_bed}}().}

\item{nCores}{integer. Number of cores to run the function with. Multicore
capability not available in Windows OS.}
}
\value{
data.table
}
\description{
Adds a column of characters representing the corresponsing nucleotide at
the reference genome position, considering the strand of the transcript.
The input DT must be the output, or resemble it, of the
}
