% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_refSeqDT}
\alias{tx_add_refSeqDT}
\title{Add reference sequence to a data.table}
\usage{
tx_add_refSeqDT(DT, genome, geneAnnot, nCores = 1)
}
\arguments{
\item{DT}{data.table. A data.table object. See
\code{\link{tx_makeDT_coverage}}, \code{\link{tx_makeDT_nucFreq}} or
\code{\link{tx_makeDT_covNucFreq}} functions.}

\item{genome}{list. The full reference genome sequences, as prepackaged
by BSgenome, See ?BSgenome::available.genomes(); or loaded by \code{\link{tx_load_genome}}}

\item{geneAnnot}{GRanges. Gene annotation loaded as a GenomicRanges object,
see tx_load_bed().}

\item{nCores}{integer. Number of cores to use to run function. Multicore
capability not available in Windows OS.}
}
\value{
data.table
}
\description{
Adds a column of characters representing the corresponsing nucleotide at
the reference genome position, considering the strand of the transcript.
The input DT must be the output, or resemble it, of the
}
\author{
M.A. Garcia-Campos
}
