% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_relTxPos}
\alias{tx_add_relTxPos}
\title{Add relative position in transcript}
\usage{
tx_add_relTxPos(txDT, round_dig = 3)
}
\arguments{
\item{txDT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{round_dig}{}
}
\value{
data.table
}
\description{
Adds a numeric variable which represents the relative position of the
nucleotide along the transcript in a scale of 0 (start) to 1 (end).
}
