% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_siteAnnotation}
\alias{tx_add_siteAnnotation}
\title{Add 1bp-site logical annotation}
\usage{
tx_add_siteAnnotation(DT, GRanges, colName, nCores = 1)
}
\arguments{
\item{DT}{data.table. Output of txtools data.tables with coverage information
as output of tx_coverageDT() and tx_covNucFreqDT()}

\item{GRanges}{GRanges. Ranges of length 1, to be marked in the data.table}

\item{colName}{character. Name of the new column to be added.}

\item{nCores}{integer. Number of cores to use to run function. Multicore
capability not available in Windows OS.}
}
\value{
data.table
}
\description{
Add a logical variable column in DT, for which the genomic coordinates of a
GRanges object are used to mark differentiate between sites in the GRanges
object and all others. TRUE = site in GRanges object. This is useful for
example when marking already known RNA modification sites.
}
