% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_siteAnnotation}
\alias{tx_add_siteAnnotation}
\title{Add site annotation}
\usage{
tx_add_siteAnnotation(x, GR, type = "logical", colName)
}
\arguments{
\item{x}{data.table. Output of txtools data.tables with coverage information
as output of tx_coverageDT() and tx_covNucFreqDT()}

\item{GR}{GenomicRanges. Length 1 ranges which want to be marked in the
data.tables objects.}

\item{type}{character. Type of variable to be added:
1) 'logical': Found coordinates will be marked as TRUE while the rest will be
left as FALSE.}

\item{colName}{character. Name of the new column to be added.}
}
\value{

}
\description{
Add a logical variable column in which desired genomic coordinates are
marked as TRUE. Useful for example when distinguishing between known
modified sites.
}
