% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_add_startRatio}
\alias{tx_add_startRatio}
\title{Add startRatio to data.table}
\usage{
tx_add_startRatio(DT)
}
\arguments{
\item{DT}{data.table. Output of txtools data.tables with coverage information
as output of \code{\link{tx_coverageDT}} \code{\link{tx_covNucFreqDT}}
functions}
}
\value{
data.table
}
\description{
Add read-starts ratio over coverage.
}
