% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_complete_DT}
\alias{tx_complete_DT}
\title{Complete a DT object missing genes}
\usage{
tx_complete_DT(DT, geneAnnot, genome = NULL, nCores = 1)
}
\arguments{
\item{DT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{geneAnnot}{GRanges. Gene annotation as loaded by \code{\link{tx_load_bed}}().}

\item{genome}{list. The full reference genome sequences, as loaded by
\code{\link{tx_load_genome}}() or prepackaged by BSgenome, see ?BSgenome::available.genomes}

\item{nCores}{integer. Number of cores to run the function with. Multicore
capability not available in Windows OS.}
}
\value{
data.table
}
\description{
A function that adds the genomic and transcriptomic coordinates of genes
with no data in a data.table generated by the \code{\link{tx_coverageDT}},
\code{\link{tx_nucFreqDT}}, and \code{\link{tx_covNucFreqDT}} functions.
}
\details{
This function is meant to equalize datasets which don't contain some genes,
as such the order of the genes is the same as in the GRanges object used as
gene annotation.
}
