% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_complete_DT}
\alias{tx_complete_DT}
\title{Complete and sort a DT according to a gene annotation in GRanges format}
\usage{
tx_complete_DT(DT, geneAnnot, fastaGenome = NULL, nCores = 1)
}
\arguments{
\item{DT}{data.table .}

\item{geneAnnot}{GRanges}

\item{fastaGenome}{character}

\item{nCores}{numeric}
}
\value{

}
\description{
Complete and sort a DT according to a gene annotation in GRanges format
}
