% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_counts}
\alias{tx_counts}
\title{Total counts of reads per gene model}
\usage{
tx_counts(x)
}
\arguments{
\item{x}{CompressedGRangesList. Genomic Ranges list containing genomic
alignments data by gene. Constructed via the tx_reads() function.}
}
\value{
integer
}
\description{
Total counts of reads per gene model
}
