% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_counts}
\alias{tx_counts}
\title{Quantifies reads by gene model from a tx_reads list}
\usage{
tx_counts(x)
}
\arguments{
\item{x}{GenomicRanges.}
}
\value{

}
\description{
Quantifies reads by gene model from a tx_reads list
}
\author{
M.A. Garcia-Campos
}
