% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_coverageDT}
\alias{tx_coverageDT}
\title{Summarized Coverage data.table}
\usage{
tx_coverageDT(x, geneAnnot, nCores = 1)
}
\arguments{
\item{x}{CompressedGRangesList. Genomic Ranges list containing genomic
alignments data by gene. Constructed via the tx_reads() or tx_reads_mc()
functions.}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed()}

\item{nCores}{integer. Number of cores to use to run function.}
}
\value{
data.table
}
\description{
Summarized Coverage data.table
}
\author{
M.A. Garcia-Campos
}
