% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_coverageDT}
\alias{tx_coverageDT}
\title{Summarized Coverage data.table}
\usage{
tx_coverageDT(x, geneAnnot)
}
\arguments{
\item{geneAnnot}{}
}
\value{
data.table
}
\description{
Summarized Coverage data.table
}
