% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_coverageDT}
\alias{tx_coverageDT}
\title{Summarized Coverage data.table}
\usage{
tx_coverageDT(x, geneAnnot, nCores = 1)
}
\arguments{
\item{x}{CompressedGRangesList. Genomic Ranges list containing genomic
alignments data by gene. Constructed via the tx_reads() or tx_reads_mc()
functions.}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed()}

\item{nCores}{integer. Number of cores to use to run function.}
}
\value{
data.table
}
\description{
This function constructs a list of data.tables that contains coverage metrics
per nucleotide in transcript:
\itemize{
\item cov = Insert coverage
\item start_5p = read-start counts
\item end_3p = read-end counts
}
The function requires the input of a GRangesList object output by the
\code{\link{tx_reads}} function, which should contain sequence alignments in the
transcriptomic space, and a gene annotation in GRanges format, as loaded by
the \code{\link{tx_load_bed}} function.
}
\details{
This function allows for usage of multiple cores to reduce processing times
in UNIX-like OS.
}
\author{
M.A. Garcia-Campos
}
