% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_coverageTab}
\alias{tx_coverageTab}
\title{Calculate coverage table: coverage, 5prime-starts, and 3prime-ends}
\usage{
tx_coverageTab(x)
}
\arguments{
\item{x}{CompressedGRangesList}
}
\value{
data.table
}
\description{
Calculate coverage table: coverage, 5prime-starts, and 3prime-ends
}
\author{
M.A. Garcia-Campos
}
