% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_coverageTab_mc}
\alias{tx_coverageTab_mc}
\title{Calculate coverage table (cov, ends, starts) for all gene models (Multicore)}
\usage{
tx_coverageTab_mc(x, nCores)
}
\arguments{
\item{nCores}{}
}
\value{

}
\description{
Calculate coverage table (cov, ends, starts) for all gene models (Multicore)
}
