% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_covNucFreqDT}
\alias{tx_covNucFreqDT}
\title{Summarized Coverage & Nucleotide Frequency data.table}
\usage{
tx_covNucFreqDT(x, geneAnnot, simplify_IUPAC = "splitForceInt", nCores = 1)
}
\arguments{
\item{x}{CompressedGRangesList. Genomic Ranges list containing genomic
alignments data by gene. Constructed via the tx_reads() or tx_reads_mc()
functions.}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed()
function.}

\item{simplify_IUPAC}{string. Available options are :
\itemize{
\item "not": Will output the complete nucleotide frequency table including
ambiguous reads using the IUPAC ambiguity code. Check: ?Biostrings::IUPAC_CODE_MAP
\item "splitForceInt": Will force an integers split in which ambiguous codes
will be split and assigned half the frequency into their respective nucleotides,
if the frequency is an odd number the uneven count will be assigned as "N".
\item "splitHalf": Ambiguous nucleotide frequencies will be split in half to
their corresponding nucleotides, in cases where frequency is odd creating
non-integer frequencies.
}}

\item{nCores}{integer. Number of cores to use to run function.}
}
\value{
data.table
}
\description{
#' This function constructs a list of data.tables that contains nucleotide frequency
metrics per nucleotide by transcript:
\itemize{
\item A = Adenine
\item C = Citocine
\item G = Guanine
\item T = Timine
\item N = Read but undetermined
\item - = Deletion
\item . = Gap between read1 and read2
}
The function requires the input of a GRangesList object output by the
\code{\link{tx_reads}} function, which should contain sequence alignments in the
transcriptomic space, and a gene annotation in GRanges format, as loaded by
the \code{\link{tx_load_bed}} function.
}
\details{
This function allows for usage of multiple cores to reduce processing times
in UNIX-like OS.
}
\author{
M.A. Garcia-Campos
}
