% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_cut_geneAnnotBytxDT}
\alias{tx_cut_geneAnnotBytxDT}
\title{Cut gene annotation by txDT's genes}
\usage{
tx_cut_geneAnnotBytxDT(txDT, geneAnnot)
}
\arguments{
\item{txDT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{geneAnnot}{GRanges. Gene annotation as loaded by \code{\link{tx_load_bed}}()}
}
\value{
data.table
}
\description{
Cut gene annotation by txDT's genes
}
