% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_filter_maxWidth}
\alias{tx_filter_maxWidth}
\alias{tx_filter_max_width}
\title{Filter ranges by a maximum width}
\usage{
tx_filter_maxWidth(x, thr, nCores = 1)
}
\arguments{
\item{x}{CompressedGRangesList. Genomic Ranges list containing genomic
alignments data by gene. Constructed via the tx_reads() or tx_reads_mc()
functions.}

\item{thr}{integer. Threshold for maximum width size allowed on output.}

\item{nCores}{integer. Number of cores to use to run function. Multicore
capability is not available in Windows OS.}
}
\value{
CompressedGRangesList
}
\description{
Filter ranges by a maximum width
}
\author{
M.A. Garcia-Campos
}
