% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_genCoorTab}
\alias{tx_genCoorTab}
\title{Table with genomic and transcriptomic coordinates}
\usage{
tx_genCoorTab(x, geneAnnot_GR)
}
\arguments{
\item{geneAnnot_GR}{}
}
\value{

}
\description{
Table with genomic and transcriptomic coordinates
}
