% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tx_generateSingleEndFASTQ}
\alias{tx_generateSingleEndFASTQ}
\title{Generate single-end FASTQ file}
\usage{
tx_generateSingleEndFASTQ(
  genome,
  geneAnnot,
  readLen,
  libSize,
  fileName,
  NB_r = 5,
  NB_mu = 500,
  nCores
)
}
\arguments{
\item{nCores}{}
}
\value{

}
\description{
Generate single-end FASTQ file form genome and gene annotation.
Distribution of reads is randomly selected following a negative binomial distribution with
}
