% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_get_flankSequence}
\alias{tx_get_flankSequence}
\title{Get flanking sequences}
\usage{
tx_get_flankSequence(DT, logi_col, upFlank, doFlank, addNames = FALSE)
}
\arguments{
\item{addNames}{}
}
\description{
Get flanking sequences
}
