% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_get_flanksFromLogicAnnot}
\alias{tx_get_flanksFromLogicAnnot}
\title{Get data from a position and their neighboring positions}
\usage{
tx_get_flanksFromLogicAnnot(
  DT,
  logi_col,
  values_col,
  upFlank,
  doFlank,
  addRowNames = FALSE
)
}
\arguments{
\item{DT}{data.table. Output of txtools data.tables with coverage information
as output of \code{\link{tx_coverageDT}} \code{\link{tx_covNucFreqDT}}}

\item{addRowNames}{functions.}
}
\value{
matrix
}
\description{
Get the data from positions marked with TRUE in a logical variable as the
center and the values in its vicinity, delimited by a flank size.
Useful to aggregate data surrounding specific sites as motif locations.
}
