% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_get_flanksFromLogicAnnot}
\alias{tx_get_flanksFromLogicAnnot}
\title{Get data from a position and their neighboring positions}
\usage{
tx_get_flanksFromLogicAnnot(
  DT,
  logi_col,
  values_col,
  upFlank,
  doFlank,
  addRowNames = TRUE
)
}
\arguments{
\item{DT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{logi_col}{character. Name of column of logical class, which indicates queried sites}

\item{values_col}{character}

\item{upFlank}{numeric. Up-stream flank length}

\item{doFlank}{numeric. Down-stream flank length}

\item{addRowNames}{logical. Set to TRUE to add rownames in format "gene:txcoor".
Default is FALSE.}
}
\value{
matrix
}
\description{
Get the data from positions marked with TRUE in a logical variable as the
center and the values in its vicinity, delimited by a flank size.
Useful to aggregate data surrounding specific sites as motif locations.
}
