% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_get_geneLengths}
\alias{tx_get_geneLengths}
\title{Get length of genes}
\usage{
tx_get_geneLengths(DT)
}
\arguments{
\item{DT}{data.table. Output of txtools data.tables with coverage information
as output of \code{\link{tx_coverageDT}} \code{\link{tx_covNucFreqDT}}
functions.}
}
\value{
named numeric. Length of genes as per times they appear in DT
}
\description{
Outputs the length of the transcripts in the DT
}
