% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_get_metageneAtCDS}
\alias{tx_get_metageneAtCDS}
\title{Get metagene at CDS}
\usage{
tx_get_metageneAtCDS(
  txDT,
  geneAnnotation,
  colVars,
  CDS_align,
  upFlank,
  doFlank
)
}
\arguments{
\item{doFlank}{}
}
\description{
Get metagene at CDS
}
