% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_get_metageneRegions}
\alias{tx_get_metageneRegions}
\title{Get metagene regions}
\usage{
tx_get_metageneRegions(
  txDT,
  geneAnnot,
  colVars,
  nBins_5UTR,
  nBins_CDS = NULL,
  nBins_3UTR = NULL,
  nCores = 1
)
}
\arguments{
\item{txDT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{geneAnnot}{GRanges. Gene annotation as loaded by \code{\link{tx_load_bed}}().}

\item{colVars}{character. Names of columns for which values will be extracted}

\item{nBins_5UTR}{integer. Number of bins into which allocate data on 5'UTR regions}

\item{nBins_CDS}{integer. Number of bins into which allocate data on CDS regions}

\item{nBins_3UTR}{integer. Number of bins into which allocate data on 3'UTR regions}

\item{nCores}{integer. Number of cores to run the function with. Multicore
capability not available in Windows OS.}
}
\value{
list of matrices for each colVar
}
\description{
Summarizes the values of a column(s) into a selected number of bins per
gene structure region (5'UTR, CDS, 3'UTR).
}
