% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_get_transcriptSeqs}
\alias{tx_get_transcriptSeqs}
\title{Get transcriptome sequences}
\usage{
tx_get_transcriptSeqs(genome, geneAnnot, outFile = NULL, nCores = 1)
}
\arguments{
\item{genome}{list. The full reference genome sequences, as loaded by
\code{\link{tx_load_genome}}() or prepackaged by BSgenome, see ?BSgenome::available.genomes}

\item{geneAnnot}{GRanges. Gene annotation as loaded by \code{\link{tx_load_bed}}().}

\item{outFile}{character. If specified, sequences will be written to this file name.
Otherwise a character vector will be output}

\item{nCores}{integer. Number of cores to run the function with. Multi-core
capability not available in Windows OS.}
}
\value{
character
}
\description{
Get transcriptome sequences
}
