% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_load_bam}
\alias{tx_load_bam}
\title{Read paired end bam file by yield size}
\usage{
tx_load_bam(
  file,
  pairedEnd,
  yieldSize = 1e+05,
  scanFlag = "default",
  loadSeq = FALSE,
  strandMode = 1,
  verbose = TRUE
)
}
\arguments{
\item{file}{character. Path to the file to read}

\item{pairedEnd}{logical. Set to FALSE if reads in BAM file are single-end,
set to TRUE if reads are paired-end.}

\item{yieldSize}{numeric. Number of reads to be processed at a time}

\item{scanFlag}{integer. Flag used to filter reads.
See \code{\link[Rsamtools]{ScanBamParam}}}

\item{loadSeq}{logical. Set to TRUE for loading the sequences contained
in the BAM file}

\item{strandMode}{numeric. \itemize{
\item 1 (default): Strand of the pair is that of its \strong{first} alignment: Directional Illumina (Ligation), Standard SOLiD. (Single-end No change in strand)
\item 2: strand of the pair is strand of its \strong{last} alignment: dUTP, NSR, NNSR, Illumina stranded TruSeq PE protocol. (Single-end: Change to inverse strand)
\item 0: strand of the pair is set to '*' (unspecified), works as both "+" and "-".} More info: \code{\link[GenomicAlignments]{GAlignmentPairs-class}}.}
}
\value{
GRanges
}
\description{
Reads a file in BAM format by blocks of lines equal to a yield size, either
automatically calculated or specified by the user, and loads it as a
GenomicAlignments object.
}
\examples{
# Loading in-package BAM file
bamFile <- system.file("extdata", "example_hg19.bam", package = "txtools")
hg19_bam <- tx_load_bam(bamFile, pairedEnd = TRUE, loadSeq = TRUE, verbose = TRUE)
summary(hg19_bam)
}
