% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_load_genome}
\alias{tx_load_genome}
\title{Load genome}
\usage{
tx_load_genome(fastaFile)
}
\arguments{
\item{fastaFile}{path to FASTA format file with all genome sequences}
}
\value{
DNAStringSet
}
\description{
Load genome as DNAStrinSet from FASTA file. Alias for
\code{\link[Biostrings]{readDNAStringSet}}
}
