% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_makeDT_coverage}
\alias{tx_makeDT_coverage}
\alias{tx_coverageDT}
\title{Summarized Coverage data.table}
\usage{
tx_makeDT_coverage(x, geneAnnot, genome = NULL, fullDT = FALSE, nCores = 1)
}
\arguments{
\item{x}{CompressedGRangesList. Genomic Ranges list containing genomic
alignments data by gene. Constructed via the \code{\link{tx_reads}}
function.}

\item{geneAnnot}{GRanges. Gene annotation as loaded by \code{\link{tx_load_bed}}()}

\item{genome}{list. The full reference genome sequences, as loaded by
\code{\link{tx_load_genome}}() or prepackaged by BSgenome, see ?BSgenome::available.genomes}

\item{fullDT}{logical. Set to TRUE if it is desired to output a data.table
with all genes and in the same order as 'geneAnnot' object.}

\item{nCores}{integer. Number of cores to run the function with. Multicore
capability not available in Windows OS.}
}
\value{
data.table
}
\description{
This function constructs a data.table that contains coverage metrics
per nucleotide for all transcripts with corresponding GRanges in its first argument 'x':
\itemize{
\item cov = Insert coverage
\item start_5p = read-start counts
\item end_3p = read-end counts
}
}
\seealso{
Other makeDT functions: 
\code{\link{tx_makeDT_covNucFreq}()},
\code{\link{tx_makeDT_nucFreq}()}
}
\author{
M.A. Garcia-Campos
}
\concept{makeDT functions}
