% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_nucFreqTab}
\alias{tx_nucFreqTab}
\title{Calculate nucleotide frequency pileup for all gene models}
\usage{
tx_nucFreqTab(x, simplify_IUPAC = "not")
}
\arguments{
\item{x}{CompressedGRangesList. Genomic Ranges list containing genomic
alignments data by gene. Constructed via the tx_reads() or tx_reads_mc()
functions.}

\item{simplify_IUPAC}{character. Method to simplify ambiguous reads:
1) 'not': Ambiguous reads will be left as their IUPAC_ambiguous code, e.g.
for positions in which a 'G' and and 'A' where read an 'R' will note this.
2) "splitHalf': Ambiguous reads will be divided in half, in odd cases having
fractions of reads assigned.
3) "splitForceInt": Ambiguous reads will be divided in half, but forced to be
integers, unassigned fraction of reads will be summed and assigned as 'N'.}
}
\value{
data.table
}
\description{
Calculate nucleotide frequency pileup for all gene models
}
\author{
M.A. Garcia-Campos
}
