% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_plots.R
\name{tx_plot_ggseqlogo}
\alias{tx_plot_ggseqlogo}
\title{Plot motif centered in logical annotation}
\usage{
tx_plot_ggseqlogo(DT, logi_col, upFlank, doFlank, method = "bits")
}
\arguments{
\item{DT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{logi_col}{character. Name of column of logical class, which indicates queried sites}

\item{upFlank}{numeric. Up-stream flank length}

\item{doFlank}{numeric. Down-stream flank length}

\item{method}{character. Height method, can be either "bits" or "probability". Default = "bits".}
}
\value{
ggplot
}
\description{
Plots a motif of the sequence surrounding sites marked as TRUE in a logical
vector in a
}
