% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_plots.R
\name{tx_plot_metageneAtCDS}
\alias{tx_plot_metageneAtCDS}
\title{Plot metagene at CDS}
\usage{
tx_plot_metageneAtCDS(
  txDT,
  geneAnnot,
  colVars,
  CDS_align,
  upFlank,
  doFlank,
  summ_fun = "mean",
  roll_fun = NULL,
  roll_n = 100,
  roll_align = "center",
  roll_fill = NA,
  smooth = TRUE,
  spar = 0.3,
  na.rm = TRUE,
  normalize = FALSE,
  tick_by = NULL
)
}
\arguments{
\item{txDT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{geneAnnot}{GRanges. Gene annotation as loaded by \code{\link{tx_load_bed}}().}

\item{colVars}{character. Names of columns for which values will be extracted}

\item{CDS_align}{character. Either "start", "end", or "spliceSite" depending on the desired
alignment, either to CDS start, CDS end, or splicing sites, respectively.}

\item{upFlank}{numeric. Up-stream flank length}

\item{doFlank}{numeric. Down-stream flank length}

\item{summ_fun}{character. Summing function either "sum" or "mean". Default: "mean"}

\item{roll_fun}{character. Rolling function either "sum" or "mean". Default: NULL}

\item{roll_n}{numeric. Window size for rolling function}

\item{roll_align}{character. Either "center" (default), "left" or "right"}

\item{roll_fill}{vector. Either an empty vector (no fill), or a vector
(recycled to) length 3 giving left, middle and right fills. NA by default}

\item{smooth}{logical. Set to FALSE for not smoothing line.}

\item{spar}{numeric. Smoothing parameter, typically (but not necessarily) in (0,1].}

\item{na.rm}{logical. Omit all NAs from computations. Default: TRUE}

\item{normalize}{logical. If set to TRUE, values are normalized so that the
area bellow the curve approximates to 1 for each variable.}

\item{tick_by}{numeric. Distance between ticks in plot. Default: upFlank/2}
}
\value{
ggplot
}
\description{
Plot metagene at CDS
}
\seealso{
Smoothing function: \code{\link[stats]{smooth.spline}}().
}
