% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_plots.R
\name{tx_plot_metageneAtCDS}
\alias{tx_plot_metageneAtCDS}
\title{Plot metagene at CDS}
\usage{
tx_plot_metageneAtCDS(
  txDT,
  geneAnnotation,
  colVars,
  CDS_align,
  upFlank,
  doFlank,
  summ_fun = "mean",
  roll_fun = "mean",
  roll_n = 100,
  roll_align = "center",
  roll_fill = NA,
  smooth = TRUE,
  spar = 0.3,
  na.rm = TRUE,
  normalize = TRUE,
  tick_by = NULL
)
}
\arguments{
\item{txDT}{data.table. As generated with tx_makeDT_*() functions.}

\item{geneAnnotation}{GRanges. Gene annotation as loaded with tx_load_bed()}

\item{colVars}{character. Names of columns to be displayed in metagene plot}

\item{CDS_align}{character. Either "start", "end", or "spliceSite" depending on the desired
alignment to CDS start, CDS end, or splicing sites, respectively.}

\item{upFlank}{numeric. Up-stream flank}

\item{doFlank}{numeric. Down-stream flank}

\item{summ_fun}{character. Summing function either "sum" or "mean". Default: "sum"}

\item{roll_fun}{character. Rolling function either "sum" or "mean". Default: "sum"}

\item{roll_n}{numeric. Window size for rolling function}

\item{roll_align}{character. Either "center" (default), "left" or "right"}

\item{roll_fill}{vector. Either an empty vector (no fill), or a vector
(recycled to) length 3 giving left, middle and right fills. NA by default}

\item{smooth}{logical. Set to FALSE for not smoothing with spline}

\item{spar}{numeric. Smoothing parameter, typically (but not necessarily) in (0,1].
Check stats::smooth.spline}

\item{na.rm}{logical. Omit all NAs from computations. Default: TRUE}

\item{normalize}{logical. Makes all areas under the curve equivalent to 100. Default: TRUE}

\item{tick_by}{numeric. Distance between ticks in plot. Default: upFlank/2}
}
\value{
ggplot
}
\description{
Plot metagene at CDS
}
