% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_plots.R
\name{tx_plot_metageneRegions}
\alias{tx_plot_metageneRegions}
\title{Plot metagene by regions}
\usage{
tx_plot_metageneRegions(
  txDT,
  geneAnnot,
  colVars,
  nBins_5UTR,
  nBins_CDS = NULL,
  nBins_3UTR = NULL,
  summ_fun = "mean",
  smooth = TRUE,
  spar = 0.3,
  nCores = 1,
  plot_type = "lineplot"
)
}
\arguments{
\item{txDT}{data.table. As generated with tx_makeDT_*() functions.}

\item{geneAnnot}{GRanges. Gene annotation as loaded with tx_load_bed()}

\item{colVars}{character. Names of columns to be displayed in metagene plot}

\item{smooth}{logical. Set to FALSE for not smoothing with spline}

\item{spar}{numeric. Smoothing parameter, typically (but not necessarily) in (0,1].
Check stats::smooth.spline}

\item{nCores}{integer. Number of cores to use to run function.}

\item{plot_type}{character. Type of plot to be output, either "lineplot" or
"boxplot".}
}
\value{
ggplot
}
\description{
Plot metagene by regions
}
