% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_reads}
\alias{tx_reads}
\title{Transcriptomic reads convertion}
\usage{
tx_reads(reads, geneAnnot, minReads = 50, withSeq = F, verbose = T, nCores = 1)
}
\arguments{
\item{reads}{GAlignments or GAlignmentPairs. Genomic alignments to be processed}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed()}

\item{minReads}{integer. Minimum number of reads required to overlap a gene}

\item{withSeq}{logical. Set to TRUE if sequence should be preserved; 'reads'
object should contain sequences.}

\item{verbose}{logical. Set to FALSE to show less information.}

\item{nCores}{integer. Number of cores to use to run function. Multicore
capability not available in Windows OS.}
}
\value{
GRanges
space. Alignments are located in genes, instead of chromosomes.
}
\description{
Assign aligned reads to their respective gene models and convert their
positions into a transcriptomic coordinate system. It also stitches together
paired-end aligned reads into a single character 'word' in which dots '.'
separate Read1 and Read2 by their corresponding insert.
}
