% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_reads}
\alias{tx_reads}
\title{Merge paired-end reads and assign them to gene models}
\usage{
tx_reads(
  reads,
  bedR,
  overlapType = "within",
  minReads = 50,
  withSeq = F,
  verbose = T
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Merge paired-end reads and assign them to gene models
}
\author{
Miguel Angel Garcia-Campos
}
