% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_reads_mc}
\alias{tx_reads_mc}
\title{Merge paired-end reads and assign them to gene models (Multicore)}
\usage{
tx_reads_mc(
  reads,
  geneAnnot,
  nCores,
  overlapType = "within",
  minReads = 50,
  withSeq = F,
  verbose = T
)
}
\arguments{
\item{reads}{GAlignmentPairs. Paired end genomic alignments to be processed}

\item{geneAnnot}{GenomicRanges. Gene annotation loaded via the tx_load_bed()}

\item{nCores}{integer. Number of cores to be used.}

\item{overlapType}{character. Overlap type to filter reads by gene
model.}

\item{minReads}{integer. Minimum number of reads required to overlap a gene
model to be part of the output object.}

\item{verbose}{logical. Set to FALSE to show less information.}
}
\value{

}
\description{
Merge paired-end reads and assign them to gene models (Multicore)
}
\author{
M.A. Garcia-Campos
}
