% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_reads_mc}
\alias{tx_reads_mc}
\title{Merge paired-end reads and assign them to gene models (Multicore)}
\usage{
tx_reads_mc(
  reads,
  bedR,
  nCores,
  overlapType = "within",
  minReads = 50,
  withSeq = F,
  verbose = T
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Merge paired-end reads and assign them to gene models (Multicore)
}
\author{
M.A. Garcia-Campos
}
