% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_sampleByGenes}
\alias{tx_sampleByGenes}
\title{Sample txDT by genes}
\usage{
tx_sampleByGenes(txDT, size)
}
\arguments{
\item{txDT}{data.table A data.table as generated by the \code{\link{tx_coverageDT}},
\code{\link{tx_nucFreqDT}}, and \code{\link{tx_covNucFreqDT}} functions.}

\item{size}{number of genes to be sampled}
}
\value{
data.table
}
\description{
Sample txDT by genes
}
