% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_shift_geneWise}
\alias{tx_shift_geneWise}
\title{shift column in txDT}
\usage{
tx_shift_geneWise(DT, colToShift, direction, bp, nCores = 1)
}
\arguments{
\item{nCores}{}
}
\value{

}
\description{
shift column in txDT
}
