% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_shift_geneWise}
\alias{tx_shift_geneWise}
\title{Shift column in txDT}
\usage{
tx_shift_geneWise(DT, colToShift, direction, bp, nCores = 1)
}
\arguments{
\item{DT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{nCores}{}
}
\value{
data.table
}
\description{
Shift column in txDT
}
