% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_split_DT}
\alias{tx_split_DT}
\title{Split data.table to list of data.tables}
\usage{
tx_split_DT(DT, dropEmpty = TRUE)
}
\arguments{
\item{DT}{data.table. A table as output by the \code{\link{tx_makeDT_coverage}}(),
\code{\link{tx_makeDT_nucFreq}}() or \code{\link{tx_makeDT_covNucFreq}}() functions.}

\item{dropEmpty}{logical. Drops empty list elements, which occurs when data
of genes have been entirely removed, but kept listed in the x$gene factor levels.}
}
\value{
list
}
\description{
Split data.table back to list with individual data.tables by 'gene' names
}
