% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_test_LRTedgeR}
\alias{tx_test_LRTedgeR}
\title{Likelihood Ratio Test}
\usage{
tx_test_LRTedgeR(DTL, tVar, sVar, test_groups, minTrials = 50)
}
\arguments{
\item{minTrials}{}
}
\value{
data.table
}
\description{
Likelihood Ratio Test
}
