% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_test_ttest}
\alias{tx_test_ttest}
\title{t-test in txDT list}
\usage{
tx_test_ttest(DTL, cont_var, test_groups, test_na.rm = FALSE, ...)
}
\arguments{
\item{DTL}{list. List of txDT, into which each element represents a replicate of an experimental dataset.}

\item{cont_var}{character. Name of column, specifying the continuous variable
to be used in test.}

\item{test_groups}{factor. Factor specifying the "WT" and "KO" samples.}

\item{test_na.rm}{logical. Remove NAs from tests}

\item{...}{For use by \code{\link[genefilter]{rowttests}}}
}
\value{
data.table
}
\description{
Apply a t-test per nucleotide position separating by groups over a txDT list.
At least 2 samples in each group are needed.
}
