% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\name{tx_unifyTxDTL}
\alias{tx_unifyTxDTL}
\title{Unify lists of txDTs}
\usage{
tx_unifyTxDTL(
  txDTL,
  geneAnnot = NULL,
  genome = NULL,
  type = "intersection",
  nCores = 1
)
}
\arguments{
\item{nCores}{}
}
\value{
list
}
\description{
Unify lists of txDTs
}
