% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txRanges.R
\name{txRangeToBED}
\alias{txRangeToBED}
\title{Title}
\usage{
txRangeToBED(Tgene, TrangeStart, TrangeEnd, geneAnnot)
}
\arguments{
\item{Tgene}{character}

\item{TrangeStart}{integer}

\item{TrangeEnd}{integer}

\item{geneAnnot}{dataframe}
}
\value{
dataframe
}
\description{
From transcriptomic range to genomic coordinates in BED format
}
\examples{
# Need to add examples
}
