% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_core.R
\docType{package}
\name{txtools}
\alias{txtools}
\alias{txtools-package}
\title{txtools: A package facilitating analysis of RNA modifications, structures, and interactions}
\description{
txtools enables the processing, analysis, and visualization of RNA-seq data
at the nucleotide-level resolution, seamlessly integrating alignments to
the genome with transcriptomic representation. txtools’ main inputs are
BAM files and a transcriptome annotation, and the main output is a table,
capturing mismatches,  deletions, and the number of reads beginning and
ending at each nucleotide in the transcriptomic space. txtools further
facilitates downstream visualization and analyses.
}
\details{
Most of txtools' functions start with the prefix tx_ and are grouped by
families:
}
\section{tx_load_*()}{

Load initial data as genomes (FASTA), gene annotations (BED), and mapped
reads (BAM).
}

\section{tx_add_*()}{

Add a new variable to the txDT, generally by computing a ratio or frequency.
Their output is the new txDT. e.g. \code{\link{tx_add_startRatio}}(), which
adds the start to coverage ratio; \code{\link{tx_add_motifPresence}}(),
which adds the location of RNA sequence motifs across the transcriptome
}

\section{tx_get_*()}{

Extract information from a txDT and generate an object that is NOT a txDT.
e.g. \code{\link{tx_get_metageneRegions}}() which outputs a metagene matrix
with each row representing a gene and each column a bin in one of the
codifying gene regions
}

\section{tx_plot_*()}{

Plotting functions. e.g. \code{\link{tx_plot_nucFreq}}() and
\code{\link{tx_plot_staEndCov}}(), which plot the counts of data of
nucleotide frequency, and read-starts/ends and coverage respectively.
}

\section{tx_test_*()}{

Use of the txDT objects from experimental data to do statistical tests of
metrics between groups of samples. e.g. \code{\link{tx_test_ttest}}() which
performs t-tests using a list of txDTs and a vector of the groups.
}

\seealso{
Useful links:
\itemize{
  \item \doi{10.1101/2023.08.24.554738}
}

}
\author{
\strong{Maintainer}: Miguel Angel Garcia-Campos \email{garciacampos.bioinfo@gmail.com} (\href{https://orcid.org/0000-0001-5223-8174}{ORCID}) (https://angelcampos.github.io/)

}
\keyword{Coverage}
\keyword{DataImport}
\keyword{DataRepresentation}
\keyword{Epitranscriptomics}
\keyword{RNASeq}
\keyword{SNP}
\keyword{Transcription}
